requirements.txt
pyyaml
requests
pytest
python-dotenv
pydantic>=2.0
# redis       # uncomment to enable Redis task queue backend
